ai_chatbot/requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
google-generativeai>=0.8.0
pandas>=2.0.0
plotly>=5.18.0
chainlit>=1.1.0
requests>=2.31.0
python-dotenv>=1.0.0
